Updates & decorators

An important category of updates for DiffusionMCMC.jl are:

DiffusionMCMC.MCMCDiffusionImputation — Type

MCMCDiffusionImputation <: MCMCImputation

Supertype of all imputations related to diffusions.

These are essential for imputing the unobserved parts of the path, as well as sampling the starting point. The following imputation steps have been implemented.

DiffusionMCMC.PathImputation — Type

struct PathImputation{T} <: MCMCDiffusionImputation
    ρs::Vector{Vector{T}}
    aux_laws
end

Update type. An indicator for imputation of unobserved path segments. ρs are the memory parameters for the preconditioned Crank–Nicolson scheme and aux_laws are the laws for the auxiliary diffusions.

PathImputation(ρ::T, P) where T<:Number

Base constructor v1. Initialize all ρs with the same value ρ and set auxiliary laws to P.

PathImputation(ρ::T, P) where T<:Vector

Base constructor v2. Initialize each recording with its own value of ρ and set auxiliary laws to P.