Index

DiffusionMCMC.AdaptationPathImputation

DiffusionMCMC.BlockingChequerboardSwitch

DiffusionMCMC.BlockingType

DiffusionMCMC.DiffusionGlobalSubworkspace

DiffusionMCMC.DiffusionGlobalWorkspace

DiffusionMCMC.DiffusionLocalSubworkspace

DiffusionMCMC.DiffusionLocalWorkspace

DiffusionMCMC.DiffusionMCMCBackend

DiffusionMCMC.DiffusionSpecificCallback

DiffusionMCMC.MCMCDiffusionImputation

DiffusionMCMC.PathImputation

DataStructures.reset!

DiffusionMCMC.init_paths!

ExtensibleMCMC.acceptance_rate

ExtensibleMCMC.acceptance_rate!

ExtensibleMCMC.readjust!

ExtensibleMCMC.register!

ExtensibleMCMC.time_to_update

DiffusionMCMC.AdaptationPathImputation — Type

mutable struct AdaptationPathImputation <: eMCMC.Adaptation
    proposed::Int64
    accepted::Int64
    target_accpt_rate::Float64
    adapt_every_k_steps::Int64
    scale::Float64
    min::Float64
    max::Float64
    offset::Float64
end

A struct containing information about the way in which to adapt the memory parameter of the preconditioned Crank–Nicolson scheme. proposed and accepted are the internal counters that keep track of the number of proposed and accepted samples. target_accpt_rate is the acceptance rate of the Metropolis-Hastings steps that is supposed to be targetted. min is used to enforce the minimal allowable range that the random walker can sample from, max enforces the maximum. offset introduces some delay in the start of decelerting the adaptation extent and scale is a scaling parameter for adaptation speed.

DiffusionMCMC.DiffusionMCMCBackend — Type

GenericMCMCBackend

A flag that no specific backend is passed.

DiffusionMCMC.BlockingChequerboardSwitch — Type

BlockingChequerboardSwitch <: BlockingType

A decorator indicating a switch (in a chequerboard pattern) that changes the locations of delimiters of blocks.

DiffusionMCMC.BlockingType — Type

BlockingType <: MCMCUpdateDecorator

Supertype of all blocking patterns.

DiffusionMCMC.DiffusionSpecificCallback — Type

DiffusionSpecificCallback <: Callback

Supertype of all callbacks specific to diffusion processes.

DataStructures.reset! — Method

reset!(adpt::AdaptationPathImputation)

Reset the number of proposals and accepted samples to zero.

DiffusionMCMC.init_paths! — Method

init_paths!(P, WW, Wnr, XX, data)

Sample paths of guided proposals without using the preconditioned Crank–Nicolson scheme.

ExtensibleMCMC.acceptance_rate! — Method

acceptance_rate!(adpt::AdaptationPathImputation)

Destructive computation of a current acceptance rate that also resets the number of proposals and accepted samples to zeros.

ExtensibleMCMC.acceptance_rate — Method

acceptance_rate(adpt::AdaptationPathImputation)

Compute current acceptance rate of the Metropolis-Hastings update step

ExtensibleMCMC.readjust! — Function

readjust!(
    imp::PathImputation, adpt::AdaptationPathImputation, mcmc_iter, i=1
)

Adaptive readjustment for the ith recording of the preconditioned Crank–Nicolson scheme's memory parameter.

ExtensibleMCMC.register! — Method

register!(updt, adpt::AdaptationPathImputation, accepted::Bool, ::Any)

Register the result of the acceptance decision in the Metropolis-Hastings step.

ExtensibleMCMC.time_to_update — Method

time_to_update(::Val{true}, adpt::AdaptationPathImputation)

Return true if it's the time to update the memory parameter of the preconditioned Crank–Nicolson scheme.